Rounding ops cleanup
