+++ tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_math_unary_sfpu_api.h
@@ -20,9 +20,7 @@
 #include "llk_math_eltwise_unary_sfpu_sigmoid.h"
 #include "llk_math_eltwise_unary_sfpu_sign.h"
 #include "llk_math_eltwise_unary_sfpu_signbit.h"
-#include "llk_math_eltwise_unary_sfpu_floor.h"
-#include "llk_math_eltwise_unary_sfpu_ceil.h"
-#include "llk_math_eltwise_unary_sfpu_round.h"
+#include "llk_math_eltwise_unary_sfpu_rounding_ops.h"
 #include "llk_math_eltwise_unary_sfpu_silu.h"
 #include "llk_math_eltwise_unary_sfpu_square.h"
 #include "llk_math_eltwise_unary_sfpu_tanh.h"
