+++ ttnn/cpp/ttnn/operations/eltwise/unary/common/unary_op_types.hpp
@@ -83,9 +83,7 @@ enum class UnaryOpType {
     BITWISE_OR,
     RIGHT_SHIFT,
     FLOOR,
-    FLOOR_FLOAT32,
     CEIL,
-    CEIL_FLOAT32,
     ROUND,
     LEFT_SHIFT,
     REMAINDER,
