+++ ttnn/cpp/ttnn/operations/eltwise/unary/unary.cpp
@@ -115,6 +115,8 @@ template struct ExecuteUnary<UnaryOpType::TANH>;
 template struct ExecuteUnary<UnaryOpType::TILED_PROD>;
 template struct ExecuteUnary<UnaryOpType::BITWISE_NOT>;
 template struct ExecuteUnary<UnaryOpType::ALT_COMPLEX_ROTATE90>;
+template struct ExecuteUnary<UnaryOpType::CEIL>;
+template struct ExecuteUnary<UnaryOpType::FLOOR>;
 
 template <UnaryOpType unary_op_type>
 Tensor ExecuteUnaryWithFastAndApproximateMode<unary_op_type>::invoke(
@@ -303,32 +305,6 @@ Tensor Abs::invoke(
 
 Tensor Abs::invoke(const ComplexTensor& input_tensor, const MemoryConfig& output_mem_config) {
     return ttnn::hypot(input_tensor[0], input_tensor[1], output_mem_config);
-}
-
-Tensor Floor::invoke(
-    QueueId queue_id,
-    const Tensor& input_tensor,
-    const std::optional<MemoryConfig>& memory_config,
-    const std::optional<Tensor>& optional_output_tensor) {
-    UnaryOpType op_type = UnaryOpType::FLOOR;
-    if (input_tensor.get_dtype() == DataType::FLOAT32) {
-        op_type = UnaryOpType::FLOOR_FLOAT32;
-    }
-
-    return detail::unary_impl(queue_id, input_tensor, {UnaryWithParam{op_type}}, memory_config, optional_output_tensor);
-}
-
-Tensor Ceil::invoke(
-    QueueId queue_id,
-    const Tensor& input_tensor,
-    const std::optional<MemoryConfig>& memory_config,
-    const std::optional<Tensor>& optional_output_tensor) {
-    UnaryOpType op_type = UnaryOpType::CEIL;
-    if (input_tensor.get_dtype() == DataType::FLOAT32) {
-        op_type = UnaryOpType::CEIL_FLOAT32;
-    }
-
-    return detail::unary_impl(queue_id, input_tensor, {UnaryWithParam{op_type}}, memory_config, optional_output_tensor);
 }
 
 Tensor Mish::invoke(
